update of the structure
