+++ tools/__init__.py
@@ -1,5 +1,6 @@
 from .grid import *
 from .isovist import *
 from .shortest_path import *
+from .shadow_space import *
 
 __all__ = [name for name in dir() if not name.startswith('_')]
